add "gcp" group
